+++ plugin/packages/wakatime/wakatime/__init__.py
@@ -3,7 +3,9 @@
     wakatime
     ~~~~~~~~
 
-    Action event appender for Wakati.Me, auto time tracking for text editors.
+    Common interface to WakaTime.com for most text editor plugins.
+    WakaTime.com is fully automatic time tracking for text editors.
+    More info at http://wakatime.com
 
     :copyright: (c) 2013 Alan Hamlett.
     :license: BSD, see LICENSE for more details.
@@ -12,7 +14,7 @@
 from __future__ import print_function
 
 __title__ = 'wakatime'
-__version__ = '0.4.10'
+__version__ = '0.5.0'
 __author__ = 'Alan Hamlett'
 __license__ = 'BSD'
 __copyright__ = 'Copyright 2013 Alan Hamlett'
@@ -26,6 +28,15 @@ import re
 import sys
 import time
 import traceback
+try:
+    import ConfigParser as configparser
+except ImportError:
+    import configparser
+try:
+    from urllib2 import HTTPError, Request, urlopen
+except ImportError:
+    from urllib.error import HTTPError
+    from urllib.request import Request, urlopen
 
 sys.path.insert(0, os.path.dirname(os.path.abspath(__file__)))
 sys.path.insert(0, os.path.join(os.path.dirname(os.path.abspath(__file__)), 'packages'))
@@ -35,11 +46,6 @@ from .stats import get_file_stats
 from .packages import argparse
 from .packages import simplejson as json
 from .packages import tzlocal
-try:
-    from urllib2 import HTTPError, Request, urlopen
-except ImportError:
-    from urllib.error import HTTPError
-    from urllib.request import Request, urlopen
 
 
 log = logging.getLogger(__name__)
@@ -52,46 +58,82 @@ class FileAction(argparse.Action):
         setattr(namespace, self.dest, values)
 
 
-def parseConfigFile(configFile):
-    if not configFile:
-        configFile = os.path.join(os.path.expanduser('~'), '.wakatime.conf')
+def upgradeConfigFile(configFile):
+    """For backwards-compatibility, upgrade the existing config file
+    to work with configparser and rename from .wakatime.conf to .wakatime.cfg.
+    """
 
-    # define default config values
-    configs = {
-        'api_key': None,
-        'ignore': [],
-        'verbose': False,
-    }
+    if os.path.isfile(configFile):
+        # if upgraded cfg file already exists, don't overwrite it
+        return
 
-    if not os.path.isfile(configFile):
-        return configs
-
+    oldConfig = os.path.join(os.path.expanduser('~'), '.wakatime.conf')
     try:
-        with open(configFile) as fh:
-            for line in fh:
+        configs = {
+            'ignore': [],
+        }
+
+        with open(oldConfig) as fh:
+            for line in fh.readlines():
                 line = line.split('=', 1)
                 if len(line) == 2 and line[0].strip() and line[1].strip():
-                    line[0] = line[0].strip()
-                    line[1] = line[1].strip()
-                    if line[0] in configs:
-                        if isinstance(configs[line[0]], list):
-                            configs[line[0]].append(line[1])
-                        elif isinstance(configs[line[0]], bool):
-                            configs[line[0]] = True if line[1].lower() == 'true' else False
-                        else:
-                            configs[line[0]] = line[1]
+                    if line[0].strip() == 'ignore':
+                        configs['ignore'].append(line[1].strip())
                     else:
-                        configs[line[0]] = line[1]
+                        configs[line[0].strip()] = line[1].strip()
+
+        with open(configFile, 'w') as fh:
+            fh.write("[settings]\n")
+            for name, value in configs.items():
+                if isinstance(value, list):
+                    fh.write("%s=\n" % name)
+                    for item in value:
+                        fh.write("    %s\n" % item)
+                else:
+                    fh.write("%s = %s\n" % (name, value))
+
+        os.remove(oldConfig)
     except IOError:
-        print('Error: Could not read from config file ~/.wakatime.conf')
+        pass
+
+
+def parseConfigFile(configFile):
+    """Returns a configparser.SafeConfigParser instance with configs
+    read from the config file. Default location of the config file is
+    at ~/.wakatime.cfg.
+    """
+
+    if not configFile:
+        configFile = os.path.join(os.path.expanduser('~'), '.wakatime.cfg')
+
+    upgradeConfigFile(configFile)
+
+    configs = configparser.SafeConfigParser()
+    try:
+        with open(configFile) as fh:
+            try:
+                configs.readfp(fh)
+            except configparser.Error:
+                print(traceback.format_exc())
+                return None
+    except IOError:
+        if not os.path.isfile(configFile):
+            print('Error: Could not read from config file ~/.wakatime.conf')
     return configs
 
 
 def parseArguments(argv):
+    """Parse command line arguments and configs from ~/.wakatime.cfg.
+    Command line arguments take precedence over config file settings.
+    Returns instances of ArgumentParser and SafeConfigParser.
+    """
+
     try:
         sys.argv
     except AttributeError:
         sys.argv = argv
+
+    # define supported command line arguments
     parser = argparse.ArgumentParser(
             description='Wakati.Me event api appender')
     parser.add_argument('--file', dest='targetFile', metavar='file',
@@ -119,27 +161,45 @@ def parseArguments(argv):
     parser.add_argument('--verbose', dest='verbose', action='store_true',
             help='turns on debug messages in log file')
     parser.add_argument('--version', action='version', version=__version__)
+
+    # parse command line arguments
     args = parser.parse_args(args=argv[1:])
+
+    # use current unix epoch timestamp by default
     if not args.timestamp:
         args.timestamp = time.time()
 
-    # set arguments from config file
+    # parse ~/.wakatime.cfg file
     configs = parseConfigFile(args.config)
+    if configs is None:
+        return args, configs
+
+    # update args from configs
     if not args.key:
-        default_key = configs.get('api_key')
+        default_key = None
+        if configs.has_option('settings', 'api_key'):
+            default_key = configs.get('settings', 'api_key')
         if default_key:
             args.key = default_key
         else:
             parser.error('Missing api key')
-    for pattern in configs.get('ignore', []):
-        if not args.ignore:
-            args.ignore = []
-        args.ignore.append(pattern)
-    if not args.verbose and 'verbose' in configs:
-        args.verbose = configs['verbose']
-    if not args.logfile and 'logfile' in configs:
-        args.logfile = configs['logfile']
-    return args
+    if not args.ignore:
+        args.ignore = []
+    if configs.has_option('settings', 'ignore'):
+        try:
+            for pattern in configs.get('settings', 'ignore').split("\n"):
+                if pattern.strip() != '':
+                    args.ignore.append(pattern)
+        except TypeError:
+            pass
+    if not args.verbose and configs.has_option('settings', 'verbose'):
+        args.verbose = configs.getboolean('settings', 'verbose')
+    if not args.verbose and configs.has_option('settings', 'debug'):
+        args.verbose = configs.getboolean('settings', 'debug')
+    if not args.logfile and configs.has_option('settings', 'logfile'):
+        args.logfile = configs.get('settings', 'logfile')
+
+    return args, configs
 
 
 def should_ignore(fileName, patterns):
@@ -233,28 +293,39 @@ def send_action(project=None, branch=None, stats={}, key=None, targetFile=None,
 def main(argv=None):
     if not argv:
         argv = sys.argv
-    args = parseArguments(argv)
+
+    args, configs = parseArguments(argv)
+    if configs is None:
+        return 103 # config file parsing error
+
     setup_logging(args, __version__)
+
     ignore = should_ignore(args.targetFile, args.ignore)
     if ignore is not False:
         log.debug('File ignored because matches pattern: %s' % ignore)
         return 0
+
     if os.path.isfile(args.targetFile):
-        branch = None
-        name = None
+
         stats = get_file_stats(args.targetFile)
-        project = find_project(args.targetFile)
+
+        project = find_project(args.targetFile, configs=configs)
+        branch = None
+        project_name = None
         if project:
             branch = project.branch()
-            name = project.name()
+            project_name = project.name()
+
         if send_action(
-                project=name,
+                project=project_name,
                 branch=branch,
                 stats=stats,
                 **vars(args)
             ):
-            return 0
-        return 102
+            return 0 # success
+
+        return 102 # api error
+
     else:
         log.debug('File does not exist; ignoring this action.')
-    return 101
+        return 0